Proyecto2/ModeloIA/requirements.txt
ultralytics
opencv-python
huggingface-hub
easyocr
paho-mqtt
